LF-5298 Rename DayWeatherSummary to WeatherDetail
